requirements.txt
joblib==0.16.0
-e git+https://github.com/nkmr-lab/manga109api.git@a5d888892054b1e9d914a428035fd7c2bb5e8347#egg=manga109api
mecab-python3==1.0.1
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.1.0
Pillow==7.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.48.0
xmltodict==0.12.0
